+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 * Add `-q` and `--quiet` flags to the CLI Tool and their respective command classes [#48](https://github.com/contentful/contentful-bootstrap.rb/issues/48)
+* Add `:no_input` option to library commands [#48](https://github.com/contentful/contentful-bootstrap.rb/issues/48)
 
 ### Changed
 * Refactored internals to allow more option flexibility and simplified the `CommandRunner`.